+++ BroadleafCommerceFramework/src/main/java/org/broadleafcommerce/content/service/ContentServiceImpl.java
@@ -55,36 +55,36 @@ import org.xml.sax.InputSource;
 public class ContentServiceImpl implements ContentService {
     private static final Log LOG = LogFactory.getLog(ContentServiceImpl.class);
     private static final LRUMap EXPRESSION_CACHE = new LRUMap(100);
-    
+
 
     @Resource(name="blContentDao")
 	protected ContentDao contentDao;
-	
+
 	@Resource(name="blContentDetailsDao")
 	protected ContentDetailsDao contentDetailsDao;
-	
+
 	/* (non-Javadoc)
 	 * @see org.broadleafcommerce.content.service.ContentService#findContentById(java.lang.Long)
 	 */
-	public Content findContentById(Long id) {
+	public Content findContentById(Integer id) {
 		return contentDao.readContentById(id);
 	}
 
 	/* (non-Javadoc)
 	 * @see org.broadleafcommerce.content.service.ContentService#findContentDetailsById(java.lang.Long)
 	 */
-	public Content findContentDetailsById(Long id) {
+	public Content findContentDetailsById(Integer id) {
 		return contentDao.readContentById(id);
 	}
-	
+
 	public List<ContentDetails> findContentDetails(String sandbox, String contentType, Map<String, Object> mvelParameters){
 		return findContentDetails(sandbox, contentType, mvelParameters, new Date(DateUtil.getNow()));
 	}
-	
+
 	public List<ContentDetails> findContentDetails(String sandbox, String contentType, Map<String, Object> mvelParameters, Date displayDate){
-		List<Content> contents = contentDao.readContentSpecified(sandbox, contentType, displayDate);		
-		List<Long> contentIds = new ArrayList<Long>();
-		
+		List<Content> contents = contentDao.readContentSpecified(sandbox, contentType, displayDate);
+		List<Integer> contentIds = new ArrayList<Integer>();
+
 		for (Content content : contents){
 			if(mvelParameters != null && content.getDisplayRule() != null && content.getDisplayRule() != ""){
 				if(!executeExpression(content.getDisplayRule(), mvelParameters)){
@@ -92,44 +92,44 @@ public class ContentServiceImpl implements ContentService {
 				}
 			}else{
 				contentIds.add(content.getId());
-			}			
+			}
 		}
 		if(contentIds.size() > 0){
-			return contentDetailsDao.readContentDetailsByOrderedIds(contentIds);							
+			return contentDetailsDao.readContentDetailsByOrderedIds(contentIds);
 		}else{
 			return new ArrayList<ContentDetails>();
 		}
-		
+
 	}
 
 	public String renderedContentDetails(String styleSheetString, List<ContentDetails> contentDetails) throws Exception{
 		return renderedContentDetails(styleSheetString, contentDetails, -1);
 	}
-	
+
 	public String renderedContentDetails(String styleSheetString, List<ContentDetails> contentDetails, int rowCount) throws Exception{
 		Source xmlSource;
 		int maxCount = (rowCount > -1 && contentDetails.size() > 0)? rowCount : contentDetails.size();
-		
+
 		Writer resultWriter = new StringWriter();
 	    StreamResult result = new StreamResult(resultWriter);
 	    Source styleSheetSource = getSource(styleSheetString);
 	    TransformerFactory transformerFactory = TransformerFactory.newInstance();
 	    Transformer transformer = transformerFactory.newTransformer(styleSheetSource);
-	    
+
 	    for(int i=0; i < maxCount; i++){
 	    	ContentDetails contentDetail = contentDetails.get(i);
-	    	xmlSource = getSource(contentDetail.getXmlContent());	    	
+	    	xmlSource = getSource(contentDetail.getXmlContent());
 	    	try{
-	    		transformer.transform(xmlSource, result);	    	
+	    		transformer.transform(xmlSource, result);
 	    	}catch (Exception e){
 	    		LOG.error("Error during transformation. ",e);
 	    		throw e;
 	    	}
-	    	
+
 	    }
 		return StringEscapeUtils.unescapeXml(resultWriter.toString()) ;
 	}
-	
+
     protected Boolean executeExpression(String expression, Map<String, Object> vars) {
         Serializable exp = (Serializable)EXPRESSION_CACHE.get(expression);
         if (exp == null) {
@@ -141,29 +141,29 @@ public class ContentServiceImpl implements ContentService {
         return (Boolean)MVEL.executeExpression(exp, vars);
 
     }
-    
+
     private Source getSource(String sourceString){
 	    StringReader styleSheetSourceReader = new StringReader(sourceString);
 	    InputSource inputStyleSheetSource = new InputSource(styleSheetSourceReader);
 	    return new SAXSource(inputStyleSheetSource);
-    	
+
     }
 
 	/* (non-Javadoc)
 	 * @see org.broadleafcommerce.content.service.ContentService#approveContent(java.util.List, java.lang.String, java.lang.String)
 	 */
-	public void approveContent(List<Long> contentIds, String sandboxName, String username) {
+	public void approveContent(List<Integer> contentIds, String sandboxName, String username) {
 		// deletes any content and associated contentDetail from the staging sandbox that
 	    // matches the passed in contentKey (i.e. type and filename).
 	    // Updates the sandBoxName for all matching content to null and sets the
 	    // approved by and approved date values
-		
+
 		List<Content> contentList = contentDao.readContentByIdsAndSandbox(contentIds, sandboxName);
 		List <Content> stageContentList = contentDao.readStagedContent();
-		
+
 		List<Content> deleteList = new ArrayList<Content>();
 		List<Content> saveList = new ArrayList<Content>();
-		
+
 		for (Content content:contentList) {
 			for (Content stageContent: stageContentList) {
 				if (stageContent.getFilePathName().equals(content.getFilePathName())
@@ -175,10 +175,10 @@ public class ContentServiceImpl implements ContentService {
 			content.setSandbox(null);
 			content.setApprovedBy(username);
 			content.setApprovedDate(new Date());
-			
+
 			saveList.add(content);
 		}
-		
+
 		contentDao.saveContent(saveList);
 		contentDao.delete(deleteList);
 	}
@@ -186,35 +186,35 @@ public class ContentServiceImpl implements ContentService {
 	/* (non-Javadoc)
 	 * @see org.broadleafcommerce.content.service.ContentService#checkoutContentToSandbox(java.util.List, java.lang.String)
 	 */
-	public List<Content> checkoutContentToSandbox(List<Long> contentIds, String sandboxName) {
+	public List<Content> checkoutContentToSandbox(List<Integer> contentIds, String sandboxName) {
 		// copies all matching content items where sandbox = null
 		// to the new records where sandbox = the passed in value
 		// sets the deployed_flag to false for all items in the new sandbox
-		
+
 		List<Content> contentList = contentDao.readContentByIdsAndSandbox(contentIds, null);
 		List<ContentDetails> contentDetailsList = contentDetailsDao.readContentDetailsByOrderedIds(contentIds);
-		
+
 		List<Content> newContentList = new ArrayList<Content>();
-		
+
 		for (Content content:contentList) {
 			// copy content and set deployed to false
 			Content newContent = new ContentImpl(content, sandboxName, false);
 			Content createdContent = contentDao.saveContent(newContent);
-			
+
 			newContentList.add(createdContent);
-			
+
 			// look up the corresponding contact detail
-			
+
 			for (ContentDetails contentDetails:contentDetailsList) {
 				if (contentDetails.getId().equals(content.getId())) {
 					ContentDetails newContentDetails = new ContentDetailsImpl(contentDetails, createdContent.getId());
 					contentDetailsDao.save(newContentDetails);
-					
+
 					break;
 				}
 			}
 		}
-		
+
 		return newContentList;
 	}
 
@@ -243,13 +243,13 @@ public class ContentServiceImpl implements ContentService {
 	/* (non-Javadoc)
 	 * @see org.broadleafcommerce.content.service.ContentService#rejectContent(java.util.List, java.lang.String, java.lang.String)
 	 */
-	public void rejectContent(List<Long> contentIds, String sandbox, String username) {
+	public void rejectContent(List<Integer> contentIds, String sandbox, String username) {
 	     // modifies the sandbox name for the matching content to equal the value in its submitted_by field
 	     //  (e.g. update content set sandbox = submitted_by where id in (....)
 	     // updates rejected by and rejected date
-		
+
 		List<Content> contentList = contentDao.readContentByIdsAndSandbox(contentIds, sandbox);
-		
+
 		for (Content content:contentList) {
 			content.setRejectedBy(username);
 			content.setRejectedDate(new Date());
@@ -257,17 +257,17 @@ public class ContentServiceImpl implements ContentService {
 			content.setSubmittedBy(null);
 			content.setSubmittedDate(null);
 		}
-		
+
 		contentDao.saveContent(contentList);
 	}
 
 	/* (non-Javadoc)
 	 * @see org.broadleafcommerce.content.service.ContentService#removeContentFromSandbox(java.util.List, java.lang.String)
 	 */
-	public void removeContentFromSandbox(List<Long> contentIds, String sandbox) {
+	public void removeContentFromSandbox(List<Integer> contentIds, String sandbox) {
 		// if sandbox is not null, deletes the content and associated content detail from the sandbox
 	    // otherwise, ignores the request
-		
+
 		if (sandbox != null) {
 			List<Content> contentList = contentDao.readContentByIdsAndSandbox(contentIds, sandbox);
 			contentDao.delete(contentList);
@@ -277,18 +277,18 @@ public class ContentServiceImpl implements ContentService {
 	/* (non-Javadoc)
 	 * @see org.broadleafcommerce.content.service.ContentService#submitContentFromSandbox(java.util.List, java.lang.String, java.lang.String)
 	 */
-	public void submitContentFromSandbox(List<Long> contentIds, String sandboxName, String username) {
+	public void submitContentFromSandbox(List<Integer> contentIds, String sandboxName, String username) {
 		// updates the sandboxName to AwaitingApproval_$username_$timestamp and sets
 		// the submitted by and submitted date values
-		
+
 		List<Content> contentList = contentDao.readContentByIdsAndSandbox(contentIds, sandboxName);
-		
+
 		for (Content content:contentList) {
 			content.setSubmittedBy(username);
 			content.setSubmittedDate(new Date());
 			content.setSandbox("AwaitingApproval_" + username + "_" + new Date().getTime());
 		}
-		
+
 		contentDao.saveContent(contentList);
 	}
 
@@ -297,13 +297,13 @@ public class ContentServiceImpl implements ContentService {
 	 */
 	public Content saveContent(Content content, ContentDetails contentDetails) {
 		Content contentFromDB = contentDao.saveContent(content);
-		
+
 		if (content != null) {
 			contentDetails.setId(contentFromDB.getId());
 			contentDetailsDao.save(contentDetails);
 		}
-		
+
 		return null;
 	}
-    
+
 }
