+++ BroadleafCommerceIntegrationTests/src/test/java/org/broadleafcommerce/content/service/ContentServiceTest.java
@@ -35,20 +35,20 @@ import org.testng.annotations.Test;
  *
  */
 public class ContentServiceTest extends BaseTest {
-	
+
 	@Resource
 	private ContentService contentService;
-	
+
 	@Resource
 	private ContentDao contentDao;
-	
+
 	@Resource
 	private ContentDetailsDao contentDetailsDao;
-	
-	private Long contentId;
-	private Long checkedOutContentId;
+
+	private Integer contentId;
+	private Integer checkedOutContentId;
 	private String readyForApprovalSandbox;
-	
+
 	@Test(groups = {"testContentService"})
 	public void testCanary() {
 		assert contentService != null;
@@ -60,65 +60,65 @@ public class ContentServiceTest extends BaseTest {
 	public void testSaveContent(Content content, ContentDetails contentDetails){
 		Content contentCreated = contentDao.saveContent(content);
 		assert contentCreated.getId() != null;
-		
+
 		contentDetails.setId(contentCreated.getId());
 		ContentDetails contentDetailsCreated = contentDetailsDao.save(contentDetails);
-		
+
 		Content contentFromDB = contentDao.readContentById(contentCreated.getId());
 		ContentDetails contentDetailsFromDB = contentDetailsDao.readContentDetailsById(contentCreated.getId());
-		
+
 		assert contentFromDB != null;
 		assert contentDetailsFromDB != null;
-		
+
 		contentId = contentCreated.getId();
 	}
-	
+
 	@Test(groups = {"testCheckoutContentToSandbox"}, dependsOnGroups = {"testSaveContent"})
 	@Transactional
 	@Rollback(false)
 	public void testCheckoutContentToSandbox() {
-		List<Long> contentIds = new ArrayList<Long>();
+		List<Integer> contentIds = new ArrayList<Integer>();
 		contentIds.add(contentId);
-		
+
 		List<Content> newContent = contentService.checkoutContentToSandbox(contentIds, "UserSandBox");
-		
+
 		assert newContent != null && !(newContent.isEmpty());
 		assert newContent.get(0).getId() != null;
-		
+
 		checkedOutContentId = newContent.get(0).getId();
 	}
-	
+
 	@Test(groups = {"testSubmitContent"}, dependsOnGroups = {"testCheckoutContentToSandbox"})
 	@Transactional
 	@Rollback(false)
 	public void testSubmitContent() {
 		assert checkedOutContentId != null;
 		assert checkedOutContentId != contentId;
-		
-		List<Long> contentIds = new ArrayList<Long>();
+
+		List<Integer> contentIds = new ArrayList<Integer>();
 		contentIds.add(checkedOutContentId);
-		
+
 		contentService.submitContentFromSandbox(contentIds, "UserSandBox", "NumeroUno");
-		
+
 		List<Content> awaitingApproval = contentDao.readContentAwaitingApproval();
 		assert awaitingApproval != null && !awaitingApproval.isEmpty();
-		
+
 		readyForApprovalSandbox = awaitingApproval.get(0).getSandbox();
 	}
-	
+
 	@Test(groups = {"testApproveContent"}, dependsOnGroups = {"testSubmitContent"})
 	@Transactional
 	public void testApproveContent() {
 		List<Content> awaitingApproval = contentDao.readContentAwaitingApproval();
 		assert awaitingApproval != null && !awaitingApproval.isEmpty();
-		
-		List<Long> contentIds = new ArrayList<Long>();
+
+		List<Integer> contentIds = new ArrayList<Integer>();
 		contentIds.add(checkedOutContentId);
-		
+
 		contentService.approveContent(contentIds, readyForApprovalSandbox, "NumeroUno");
 
 		awaitingApproval = contentDao.readContentAwaitingApproval();
 		assert awaitingApproval == null || awaitingApproval.isEmpty();
 	}
-	
+
 }
